requirements.txt
wxauto
requests
jieba.analyse
json
re
random
os
base64
zhipuai
threading